ai-categorizer/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
torch==2.2.0
torchvision==0.17.0
transformers==4.37.2
opencv-python-headless==4.9.0.80
